requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.2.0
pytest-env==1.1.2

# Linting & Formatting
flake8==6.1.0
black==23.7.0
isort==5.12.0
mypy==1.5.0

# Development Tools
ipython==8.14.0
python-dotenv==1.0.0
watchdog==3.0.0

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2
sphinx-autodoc-typehints==1.23.0

# Debugging
debugpy==1.6.7
flask-debugtoolbar==0.13.1

# Test data generation
factory-boy==3.3.0
Faker==19.3.0

# Security
safety==2.3.5
bandit==1.7.5

# Database
sqlalchemy-stubs==0.4

# Type checking
types-python-dateutil==2.8.19.7
typing-extensions==4.7.1

# Frontend development
nodeenv==1.8.0
